requirements.txt
streamlit
tensorflow
numpy
scikit-learn
pickle-mixin
requests
python-dotenv